bug: create_qr_code: Do not add BOM to string for QRcode
Adding a BOM breaks detection of anything that is not a string, http(s)://, mailto etc. Since there is no safe way to know if the input text is just "text", or something to be parsed that may break with BOM, remove the BOM in it's entirity, do not try to work around it with string startswith matching.